refactor: api 코드 분리 및 함수명 변경
