+++ src/pages/AuthCallback.tsx
@@ -1,4 +1,4 @@
-import { postLogin } from '@/apis/auth'
+import { getKakaoLogin } from '@/apis/auth'
 import { useEffect } from 'react'
 import { useNavigate, useSearchParams } from 'react-router-dom'
 
@@ -9,7 +9,7 @@ export default function AuthCallback() {
 
   useEffect(() => {
     const login = async (code: string) => {
-      const result = await postLogin(code)
+      const result = await getKakaoLogin(code)
 
       if (!result) {
         window.alert('로그인에 실패했습니다.')
